requirements.txt
imutils==0.5.4
numpy==1.23.5
opencv_python==4.6.0.66
Pillow==9.4.0
torch==1.13.1
torchvision==0.14.1
ultralytics==8.0.49
